Add api route for searching books
Allows for searching books via /books/search and some query parameters.
'name' matches on title and author.
'price' matches on price.
'language' matches on language code.
'isbn' matches on isbn and isbn 13.
Only one of each type of parameter may be passed in one query string.
Parameters are combined in as OR.